Handle mix of Single and Pedersen trustees in audit box
This changes the format of election_checksums: pki field is removed
(kept optional for now), and trustees_threshold is added (as an option
for now). Web_persist automatically regenerates audit_cache if the
trustees_threshold field is missing.

It also changes the format of the audit box, so may break things that
parse it (e.g. monitoring).