Put all public data in a single append-only archive
This commit introduces hash-chained events, and puts all public data
in a single append-only archive.

It drastically changes the format of the spool: all public data are
now available through this archive (named `$UUID.bel`); some of the
previous individual files (in particular `election.json`) are
simulated by the web server; others (in particular `ballots.jsons`)
are removed altogether. The `ballots` subdirectory with one file per
ballot is no more.

It also introduces spool versioning: the web server now checks the
version and refuses to start if the spool is not up-to-date. A
migration tool will be provided to convert a spool from the old format
to the new format.

It also makes releasing tally automatic when all trustees have done
their job. An embargo can be set with
`Web_persist.set_election_result_hidden`. When the embargo expires,
the server computes its partial decryption and attempts to compute the
result. It is still possible to manually force tally release in the
threshold mode, if some trustees are missing.

The command-line tool workflow also significantly changes. The
associated tests have been adapted.

Other changes:
- use trustee ids in `combine_factors`, avoiding costly checks
- factorize common code in `src/lib/v*/trustees.ml`
- simplify `election_result` (only `result` field is left)