Reorder calls to Eliom_reference in handle_email_post
Before this commit, OTP authentication always failed the first time.

My explanation is that unsetting `uuid_ref` before setting `env`
empties the scope and invalidates the cookie. Setting `env` afterwards
is done in a new state with a cookie that is not immediately sent to
the client. Calling `Eliom_reference.get` later, in the `email_post`
handler, then triggers the sending of a new cookie to the client.