Do no longer yield in the cryptographic backend
This is done by stopping using monads in the backend.

Rationale: performance in the browser. Ed25519 (using libsodium) is so
fast that yielding becomes non-negligible: ~5ms, compared to <0.5ms
for a group exponentiation. With this change, shuffling 1000 ballots
in the browser goes from ~60s to ~8s.

Monads were initially used for two reasons: programmatically document
the impurity of some functions (the ones using randoms), and for
cooperative scheduling (on both sides: server and browser). The first
one was just theoretical elegance, the second one will be compensated
(if needed) by other means (such as workers).