Add individual salts in credential derivation
Before this commit, the same salt (the uuid of the election) was used
to derive all public credentials.

After this commit, an invididual salt is generated for each
credential. The list of salts is published.

A new style of private credential is introduced, with the form
XXX-XXXX-XXX-XXXX-NNNN, where X are base58 digits, N are base10
digits, and NNNN is the index of the salt in the list.

We also increase the number of PBKDF2 iterations to 100k and its
output size to 512 bits, so that the bias introduced by reducing mod q
is negligible.

When the booth detects a new-style credential, it asks via a new
dedicated API endpoint the associated salt and public credential, and
checks that the derivation matches. This way, we are certain that the
private credential is correct and we do no longer need a checksum.