Add section in belenios-tool.html to partial decrypt buggy elections
In threshold mode, partial decryption keys generated before version
2.3 (more precisely, eb1418e6a5fba8010b4ed83ebfa8c705ab26654f) were
not properly reduced mod q and are now rejected. This commit adds a
section in belenios-tool.html to allow decryption of such buggy
elections.

The user has to paste their access code and private key, click on
compute, and copy the output and paste it in the regular partial
decryption form and submit.