Make voter address optional
The main voting functionality does not actually need an e-mail
address. For example, when using an external credential authority and
an external authentication for voters. The only place where an e-mail
address is actually required is when sending the voting confirmation.

When an external send_message service is set, we give it a chance to
handle messages where the recipient's "address" is just a username,
otherwise sending a message to a missing e-mail address always fails.