Add support for a grace period
Add an optional "timestamp" field to the user_info object returned by
authentication, which is supposed to give the login time. When casting
a ballot to a closed election, this timestamp is used to implement a
grace period.

Note that in the current implementation, the timestamp is always
absent. However, it is possible for an external authentication server
implementing Belenios Connect to set a timestamp.

/spend 5h